# Add peer label to TCP read and write stat queries