Closes #5693 

<details>
<summary> PR description before update </summary>
Hi and many apologies for taking so long with this one! I spent more time than usual trying to understand the metrics-api and what is happening behind the scenes with `linkerd viz stat`.

### What
---

> linkerd viz stat sums tcp_read_bytes_total and tcp_write_bytes_total over both peer="src" and peer="dst", for both inbound and outbound metrics. It should use peer="src" for inbound metrics, and peer="dst" for outbound.
[ #5693]

* I tried to keep the changes very small in this PR. I added some logic to both `stat_summary.go` and to `prometheus.go` in viz to add the peer label to TCP stat queries (only read/write).

* If we have an inbound query we add `peer="src"`; for outbound requests we add `peer="dst"`

### How
---

The solution feels a bit hacky to me. I thought of a few alternatives but ultimately decided to stick with this one since it was the simplest and smallest.

When we build the stat queries we first build a set of request labels. These labels are used for all of our stat queries -- including the latency quantiles. I thought adding `peer=...` as a label to all of them is not feasible so instead I added the label only to the TCP stats. If we request TCP stats, we add the additional peer label to the request labels and then instead of storing the tcp query in the `promQueries` map, we add the templated request to include the TCP labels.

At first I thought of including an additional `tcpLabels` argument to `getPrometheusMetrics(...)`. This would result in a larger change so I thought it'd be better to try without it first and see what people think. I also thought of creating the peer labels when we template out the queries in `getPrometheusMetrics(...)`. It got complicated pretty quickly since we need to first look at the request labels to figure out the direction and then we need to insert the label in the request labels set (request labels is a string in this case). Finally, there was the option of adding the peer label to all queries which again seemed unfeasible to me.

### Output
---

To test the output I used emojivoto. I wanted to see first how the logs differ (to confirm the label is included in the query) and second to see how the numbers differ now that the peer label is included.


```sh
# edges
SRC          DST        SRC_NS        DST_NS      SECURED
vote-bot     web        emojivoto     emojivoto   √      
web          emoji      emojivoto     emojivoto   √      
web          voting     emojivoto     emojivoto   √      
prometheus   emoji      linkerd-viz   emojivoto   √      
prometheus   vote-bot   linkerd-viz   emojivoto   √      
prometheus   voting     linkerd-viz   emojivoto   √      
```

I mainly tested with `web` and `vote-bot`.

**Before the change**

```sh
---

$ linkerd viz stat deploy/web -n emojivoto -o wide

NAME   MESHED   SUCCESS      RPS   LATENCY_P50   LATENCY_P95   LATENCY_P99   TCP_CONN   READ_BYTES/SEC   WRITE_BYTES/SEC
web       1/1    94.85%   2.3rps           3ms           5ms           5ms          3        4843.9B/s         5675.3B/s

request:\n\tsum(increase(tcp_read_bytes_total{deployment=\"web\", direction=\"inbound\", namespace=\"emojivoto\"}[1m])) by (namespace, deployment)"
request:\n\tsum(increase(tcp_write_bytes_total{deployment=\"web\", direction=\"inbound\", namespace=\"emojivoto\"}[1m])) by (namespace, deployment)"

---

$ linkerd viz stat deploy/vote-bot -n emojivoto --to deploy/web -o wide
NAME       MESHED   SUCCESS      RPS   LATENCY_P50   LATENCY_P95   LATENCY_P99   TCP_CONN   READ_BYTES/SEC   WRITE_BYTES/SEC
vote-bot      1/1    85.59%   2.0rps           5ms           9ms          10ms          1        4518.9B/s          153.8B/s

time="2021-03-15T13:33:04Z" level=debug msg="Query request:\n\tsum(increase(tcp_read_bytes_total{deployment=\"vote-bot\", direction=\"outbound\", dst_deployment=\"web\", dst_namespace=\"emojivoto\", namespace=\"emojivoto\"}[1m])) by (namespace, deployment)"
timrequest:\n\tsum(increase(tcp_write_bytes_total{deployment=\"vote-bot\", direction=\"outbound\", dst_deployment=\"web\", dst_namespace=\"emojivoto\", namespace=\"emojivoto\"}[1m])) by (namespace, deployment)"

---
 
$ linkerd viz stat deploy/web -n emojivoto --from deploy/vote-bot -o wide

NAME   MESHED   SUCCESS      RPS   LATENCY_P50   LATENCY_P95   LATENCY_P99   TCP_CONN   READ_BYTES/SEC   WRITE_BYTES/SEC
web       1/1    91.53%   2.0rps           5ms           9ms          10ms          1        4507.4B/s          155.8B/s

time="2021-03-15T13:35:29Z" level=debug msg="Query request:\n\tsum(increase(tcp_read_bytes_total{deployment=\"vote-bot\", direction=\"outbound\", dst_deployment=\"web\", dst_namespace=\"emojivoto\", namespace=\"emojivoto\"}[1m])) by (dst_namespace, dst_deployment)"
time="2021-03-15T13:35:29Z" level=debug msg="Query request:\n\tsum(increase(tcp_write_bytes_total{deployment=\"vote-bot\", direction=\"outbound\", dst_deployment=\"web\", dst_namespace=\"emojivoto\", namespace=\"emojivoto\"}[1m])) by (dst_namespace, dst_deployment)"

---
$ linkerd viz stat deploy/emoji -n emojivoto --from deploy/web -o wide
NAME    MESHED   SUCCESS      RPS   LATENCY_P50   LATENCY_P95   LATENCY_P99   TCP_CONN   READ_BYTES/SEC   WRITE_BYTES/SEC
emoji      1/1   100.00%   2.0rps           1ms           2ms           3ms          1        2211.5B/s          203.2B/s

time="2021-03-15T13:36:20Z" level=debug msg="Query request:\n\tsum(increase(tcp_read_bytes_total{deployment=\"web\", direction=\"outbound\", dst_deployment=\"emoji\", dst_namespace=\"emojivoto\", namespace=\"emojivoto\"}[1m])) by (dst_namespace, dst_deployment)"
time="2021-03-15T13:36:20Z" level=debug msg="Query request:\n\tsum(increase(tcp_write_bytes_total{deployment=\"web\", direction=\"outbound\", dst_deployment=\"emoji\", dst_namespace=\"emojivoto\", namespace=\"emojivoto\"}[1m])) by (dst_namespace, dst_deployment)"

---
$ linkerd viz stat deploy/voting -n emojivoto --from deploy/web -o wide
NAME     MESHED   SUCCESS      RPS   LATENCY_P50   LATENCY_P95   LATENCY_P99   TCP_CONN   READ_BYTES/SEC   WRITE_BYTES/SEC
voting      1/1    89.83%   1.0rps           1ms           1ms           1ms          1          56.2B/s           95.1B/s

time="2021-03-15T13:37:19Z" level=debug msg="Query request:\n\tsum(increase(tcp_read_bytes_total{deployment=\"web\", direction=\"outbound\", dst_deployment=\"voting\", dst_namespace=\"emojivoto\", namespace=\"emojivoto\"}[1m])) by (dst_namespace, dst_deployment)"
time="2021-03-15T13:37:19Z" level=debug msg="Query request:\n\tsum(increase(tcp_write_bytes_total{deployment=\"web\", direction=\"outbound\", dst_deployment=\"voting\", dst_namespace=\"emojivoto\", namespace=\"emojivoto\"}[1m])) by (dst_namespace, dst_deployment)"

```

**After the changes**: 

```sh
$ linkerd viz stat deploy/web -n emojivoto -o wide

# With peer=src READ_BYTES/SEC changed from 4k+ to 181 (what voting-bot typically sends)

NAME   MESHED   SUCCESS      RPS   LATENCY_P50   LATENCY_P95   LATENCY_P99   TCP_CONN   READ_BYTES/SEC   WRITE_BYTES/SEC
web       1/1    90.98%   2.2rps           9ms          89ms          98ms          3         181.4B/s         5082.2B/s

time="2021-03-15T13:53:42Z" level=debug msg="Query request:\n\tsum(increase(tcp_write_bytes_total{deployment=\"web\", direction=\"inbound\", namespace=\"emojivoto\", peer=\"src\"}[1m])) by (namespace, deployment)"
time="2021-03-15T13:53:42Z" level=debug msg="Query request:\n\tsum(increase(tcp_read_bytes_total{deployment=\"web\", direction=\"inbound\", namespace=\"emojivoto\", peer=\"src\"}[1m])) by (namespace, deployment)"


---
$ linkerd viz stat deploy/vote-bot -n emojivoto --to deploy/web -o wide

NAME       MESHED   SUCCESS      RPS   LATENCY_P50   LATENCY_P95   LATENCY_P99   TCP_CONN   READ_BYTES/SEC   WRITE_BYTES/SEC
vote-bot      1/1    94.69%   1.9rps           7ms          30ms          90ms          1        4320.9B/s          150.3B/s

time="2021-03-15T13:54:36Z" level=debug msg="Query request:\n\tsum(increase(tcp_read_bytes_total{deployment=\"vote-bot\", direction=\"outbound\", dst_deployment=\"web\", dst_namespace=\"emojivoto\", namespace=\"emojivoto\", peer=\"dst\"}[1m])) by (namespace, deployment)"
time="2021-03-15T13:54:36Z" level=debug msg="Query request:\n\tsum(increase(tcp_write_bytes_total{deployment=\"vote-bot\", direction=\"outbound\", dst_deployment=\"web\", dst_namespace=\"emojivoto\", namespace=\"emojivoto\", peer=\"dst\"}[1m])) by (namespace, deployment)"

---
$ linkerd viz stat deploy/web -n emojivoto --from deploy/vote-bot -o wide

NAME   MESHED   SUCCESS      RPS   LATENCY_P50   LATENCY_P95   LATENCY_P99   TCP_CONN   READ_BYTES/SEC   WRITE_BYTES/SEC
web       1/1    94.07%   2.0rps           5ms          10ms          10ms          1        4508.3B/s          152.9B/s

time="2021-03-15T13:55:25Z" level=debug msg="Query request:\n\tsum(increase(tcp_read_bytes_total{deployment=\"vote-bot\", direction=\"outbound\", dst_deployment=\"web\", dst_namespace=\"emojivoto\", namespace=\"emojivoto\", peer=\"dst\"}[1m])) by (dst_namespace, dst_deployment)"
time="2021-03-15T13:55:25Z" level=debug msg="Query request:\n\tsum(increase(tcp_write_bytes_total{deployment=\"vote-bot\", direction=\"outbound\", dst_deployment=\"web\", dst_namespace=\"emojivoto\", namespace=\"emojivoto\", peer=\"dst\"}[1m])) by (dst_namespace, dst_deployment)"v
---
$ linkerd viz stat deploy/emoji -n emojivoto --from deploy/web -o wide

NAME    MESHED   SUCCESS      RPS   LATENCY_P50   LATENCY_P95   LATENCY_P99   TCP_CONN   READ_BYTES/SEC   WRITE_BYTES/SEC
emoji      1/1   100.00%   2.0rps           1ms           2ms           3ms          1        2211.4B/s          203.2B/s

time="2021-03-15T13:56:16Z" level=debug msg="Query request:\n\tsum(increase(tcp_read_bytes_total{deployment=\"web\", direction=\"outbound\", dst_deployment=\"emoji\", dst_namespace=\"emojivoto\", namespace=\"emojivoto\", peer=\"dst\"}[1m])) by (dst_namespace, dst_deployment)"
time="2021-03-15T13:56:16Z" level=debug msg="Query request:\n\tsum(increase(tcp_write_bytes_total{deployment=\"web\", direction=\"outbound\", dst_deployment=\"emoji\", dst_namespace=\"emojivoto\", namespace=\"emojivoto\", peer=\"dst\"}[1m])) by (dst_namespace, dst_deployment)"

---
$ linkerd viz stat deploy/voting -n emojivoto --from deploy/web -o wide

NAME     MESHED   SUCCESS      RPS   LATENCY_P50   LATENCY_P95   LATENCY_P99   TCP_CONN   READ_BYTES/SEC   WRITE_BYTES/SEC
voting      1/1    77.97%   1.0rps           1ms           1ms           1ms          1          53.4B/s           95.3B/s
```


**To conclude** the labels to me seem to be set correctly (outbounds get peer=dst and inbounds get peer=src). 

For inbound reads it seems that the calculations have changed, i.e for `linkerd viz stat deploy/web` we go from  4843.9B/s to 180B/s+ which is the total amount of bytes `vote-bot` seems to usually send to our web service. 

However, I'm a bit confused because some numbers don't seem to add up -- for example when we do `linkerd viz stat deploy/vote-bot --to deploy/web` it shows that 4k+ bytes have been read in by vote-bot, I'm confused how and why this is happening. This is also true in the case of ` linkerd viz stat deploy/emoji  --from deploy/web`: it shows 2k bytes have been read in, since this is outbound I assume it's from the pov of `web` -- so web has read 2k and written 200 bytes, not sure why it read so many and I think I misunderstand something.

Anyway, apologies for the "read total" tangent and the delay in getting this together. Please let me know if the proposed approach is not well suited, thought in this case a minimal change would be better but more than happy to keep looking into it.
</details>

---

**Update**:  I've addressed Dennis and Alex's comments.

* Added a new test case for outbound tcp stats (this will also check whether the peer=dst label is set correctly on outbound requests).
* In `stat_summary` I've added a comment to indicate why we build a set of labels specifically for TCP stats.
* Most notably, I refactored `getPrometheus(...)` and got rid of the following params: _timeWindow_, _groupBy_ , _labels_ and _latencyQueryTemplate_ 
*  The function now accepts as parameters a context, a map[promType]string for request queries and a map[promType]string for latency (quantile) queries.
* Whenever we call getPrometheus(...) we template out both the request queries and latency queries and just pass these in as arguments.

I also created a helper function to template out the quantile queries based on a template and a set of labels; this is because I was repeating the same code throughout all of the calling functions and I thought it made sense to have it as a separate helper. I didn't do this for the request queries because I thought `sprintf` actually made more sense, it was a bit more explicit and shorter to write so I thought it didn't warrant its own helper function(s).

Last but not least there is a bit of duplication when we fetch the metrics asynchronously (we do the same thing, once for request queries and once for quantile queries). I'm not sure what would be more readable here, I extracted the code in a separate function but to me personally that didn't feel more readable and so I ended up leaving it as is.


### Tests
---

After refactoring, `linkerd viz stat` behaves the same way (I haven't checked gateways or routes).

```
$ linkerd viz stat deploy/web -n emojivoto -o wide

NAME   MESHED   SUCCESS      RPS   LATENCY_P50   LATENCY_P95   LATENCY_P99   TCP_CONN   READ_BYTES/SEC   WRITE_BYTES/SEC
web       1/1    91.91%   2.3rps           2ms           4ms           5ms          3         185.3B/s         5180.0B/s

# same value as before, latency seems to have dropped

time="2021-03-22T18:19:44Z" level=debug msg="Query request:\n\tsum(increase(tcp_write_bytes_total{deployment=\"web\", direction=\"inbound\", namespace=\"emojivoto\", peer=\"src\"}[1m])) by (namespace, deployment)"

time="2021-03-22T18:19:44Z" level=debug msg="Query request:\n\tsum(increase(tcp_read_bytes_total{deployment=\"web\", direction=\"inbound\", namespace=\"emojivoto\", peer=\"src\"}[1m])) by (namespace, deployment)"

# queries show the peer label
---

$ linkerd viz stat deploy/web -n emojivoto --from deploy/vote-bot -o wide

NAME   MESHED   SUCCESS      RPS   LATENCY_P50   LATENCY_P95   LATENCY_P99   TCP_CONN   READ_BYTES/SEC   WRITE_BYTES/SEC
web       1/1    93.16%   1.9rps           3ms           4ms           4ms          1        4503.4B/s          153.1B/s


# stats same as before except for latency which seems to have dropped a bit

time="2021-03-22T18:22:10Z" level=debug msg="Query request:\n\tsum(increase(tcp_write_bytes_total{deployment=\"vote-bot\", direction=\"outbound\", dst_deployment=\"web\", dst_namespace=\"emojivoto\", namespace=\"emojivoto\", peer=\"dst\"}[1m])) by (dst_namespace, dst_deployment)"

time="2021-03-22T18:22:10Z" level=debug msg="Query request:\n\tsum(increase(tcp_read_bytes_total{deployment=\"vote-bot\", direction=\"outbound\", dst_deployment=\"web\", dst_namespace=\"emojivoto\", namespace=\"emojivoto\", peer=\"dst\"}[1m])) by (dst_namespace, dst_deployment)"

# queries show the right label
```

**Update 2**: I'm getting a lint error for `promNumMirroredServices` -- it's not in use anywhere. I was wondering about this when I refactored the code, I haven't seen any queries made for the num of mirrored services. Do we want to add a query in as part of this change or remove the promType? Not sure how to proceed.

Signed-off-by: mateiidavid <matei.david.35@gmail.com>

<!--  Thanks for sending a pull request!

If you already have a well-structured git commit message, chances are GitHub
set the title and description of this PR to the git commit message subject and
body, respectively. If so, you may delete these instructions and submit your PR.

If this is your first time, please read our contributor guide:
https://github.com/linkerd/linkerd2/blob/main/CONTRIBUTING.md

The title and description of your Pull Request should match the git commit
subject and body, respectively. Git commit messages are structured as follows:

```
Subject

Problem

Solution

Validation

Fixes #[GitHub issue ID]

DCO Sign off
```

Example git commit message:

```
Introduce Pull Request Template

GitHub's community guidelines recommend a pull request template, the repo was
lacking one.

Introduce a `PULL_REQUEST_TEMPLATE.md` file.

Once merged, the
[Community profile checklist](https://github.com/linkerd/linkerd2/community)
should indicate the repo now provides a pull request template.

Fixes #3321

Signed-off-by: Jane Smith <jane.smith@example.com>
```

Note the git commit message subject becomes the pull request title.

For more details around git commits, see the section on Committing in our
contributor guide:
https://github.com/linkerd/linkerd2/blob/main/CONTRIBUTING.md#committing
-->