@adleong 2021-03-16
Thanks @mateiidavid!  I appreciate your instinct to keep this change small, even if it is a bit hacky.  Normally, I'd be fully onboard with that approach -- however in this case I think we've hit a critical mass of complexity in `getPrometheusMetrics` where that function is just trying to do too many different things.  (In fact, we were probably past that point even before this change 😬)

What do you think about changing `getPrometheusMetrics` to always expect queries to be fully populated?  This would mean changing the signature to no longer have callers pass in `labels, timeWindow, groupBy` and instead have those callers populate the queries themselves.  This makes sense to me since each query may have different parameters and it's the caller who best knows which query should get which parameters.

I know this change is a bit larger in scope than you had planned, but I think it would leave us in a better and more maintainable state.  Let me know how this sounds!