@mateiidavid 2021-03-17
Hey Alex, thanks for the review!

> What do you think about changing `getPrometheusMetrics` to always expect queries to be fully populated? This would mean changing the signature to no longer have callers pass in `labels, timeWindow, groupBy` and instead have those callers populate the queries themselves. This makes sense to me since each query may have different parameters and it's the caller who best knows which query should get which parameters.

Yeah that makes a lot of sense! On it 👀 .

Not a problem on my side that the change will be in larger in scope. If anything it's a good way for me to snoop around the metrics api a bit more. Will get this updated soon.